requirements.txt
python-telegram-bot==20.7
selenium==4.21.0
beautifulsoup4==4.12.3
pyTelegramBotAPI
flask
